simplifying by removing lunr logic and google logic used for discovery
