+++ lib/indexer/index.js
@@ -1,4 +1,4 @@
-var Elastic       = require('../engine/elastic'),
+var Elastic       = require('../engine'),
     mysql         = require('mysql'),
     async         = require('async'),
     request       = require('request'),
