+++ lib/server.js
@@ -5,20 +5,6 @@ var restify  = require('restify'),
     logger   = require('./logger'),
     swagger  = require('./swagger');
 
-/*var authMiddleware = function (req, resp, next) {
-  if (!req.authorization) {
-    return next(new restify.InvalidCredentialsError('Basic Authorization is required, with the API Key as the username'));
-  }
-  if (!(req.authorization.basic && req.authorization.basic.username)) {
-    return next(new restify.InvalidCredentialsError('API Key required'));
-  }
-  var apiKey = req.authorization.basic.username;
-  if (apiKey !== process.env.API_KEY) {
-    return next(new restify.InvalidCredentialsError('API Key not recognized'));
-  }
-  next();
-};*/
-
 module.exports = function () {
   var server = restify.createServer({
     name: 'directory',
@@ -69,9 +55,7 @@ module.exports = function () {
   }));
   server.get('/:location', api.get);
 
-  search.load(process.env.BADGE_STORE || (__dirname + '/../db/badges.json'), function (/*err*/) {
-    server.listen(process.env.PORT || 9000, function () {
-      console.log('%s listening at %s', server.name, server.url);
-    });
+  server.listen(process.env.PORT || 9000, function () {
+    console.log('%s listening at %s', server.name, server.url);
   });
 };
